formtemplate - rename files
